Added admins and some permissions

+++ routers/todos.py
@@ -31,31 +31,47 @@ class Todo(BaseModel):
     priority: int = Field(gt=0, lt=6)
     complete: bool
 
-@router.get("/", )
-async def read_all(user: user_dependency, db: db_dependency, status_code=status.HTTP_200_OK):
-    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
+@router.get("/", status_code=status.HTTP_200_OK)
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    return db.query(Todos).filter(Todos.owner_id == user.get("user_id")).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-    if todo_model is None:
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('user_id')).first()
+
+    if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(user:user_dependency,
+async def create_todo(user: user_dependency,
                       db: db_dependency,
                       todo: Todo):
     if user is None:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Authentication Failed')
-    todo_model = Todos(**todo.dict(), owner_id=user.get('id'))
+    print(user.get('id'))
+    todo_model = Todos(**todo.dict(), owner_id=user.get('user_id'))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def update_todo(db: db_dependency, todo: Todo, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id==todo_id).first()
+async def update_todo(
+        user: user_dependency,
+        db: db_dependency,
+        todo: Todo,
+        todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Authentication Failed')
+
+    todo_model = db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id == user.get('user_id')).first()
+
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
 
@@ -68,11 +84,17 @@ async def update_todo(db: db_dependency, todo: Todo, todo_id: int = Path(gt=0)):
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id==todo_id).first()
+async def delete_todo(
+        user: user_dependency,
+        db: db_dependency,
+        todo_id: int = Path(gt=0)):
+    todo_model = db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id==user.get('user_id')).first()
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Authentication Failed')
 
     if todo_model is None:
         raise HTTPException(status_code=404)
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('id')).delete()
     db.commit()
